requirements.txt
scapy>=2.5.0
rich>=13.0.0
pyyaml>=6.0
colorama>=0.4.6

